refactor: lint cleanup wave 7 — line-preserving auto-fixes (77 errors)
Apply line-preserving ruff auto-fixes (no line-number shifts → ledger tests stay green):

- **RUF059**: unused tuple unpack → `_var` (46)
- **RUF046**: unnecessary `int(...)` cast (18)
- **RUF021**: parenthesize `(a and b) or (c and d)` for explicit grouping (7 leftover)
- **SIM118**: drop `.keys()` from iteration (5 leftover)
- **SIM910**: remove `dict.get(k, None)` (1 leftover)

70 insertions / 70 deletions across 44 files. Ledger sweep green
(2840 passed / 9 skipped / 11190 deselected).

Skipped W191 (598 in 3 files: `terminal_feed_state.py`, `terminal_resolution_state.py`,
`streamlit_databento_volatility_screener.py`) — `ruff check` does not auto-fix W191
and `ruff format` would reflow content (line-shift risk). Defer to a separate PR.

+++ dashboard/decision_first_panel.py
@@ -54,7 +54,7 @@ def sparkline(values: Sequence[float]) -> str:
     span = hi - lo
     n = len(_SPARK_CHARS) - 1
     return "".join(
-        _SPARK_CHARS[int(round((v - lo) / span * n))] for v in values
+        _SPARK_CHARS[round((v - lo) / span * n)] for v in values
     )
